@@ -1,5 +1,5 @@
 module github.com/clovisphere/the-go-programming-language-book-exercises-solution/chap_01/ex_1_10
 
-go 1.14
+go 1.22
 
 require github.com/enescakir/emoji v1.0.0
